+++ tests/browser/select.test.js
@@ -5,7 +5,7 @@ const Wendigo = require('../../lib/wendigo');
 const configUrls = require('../config.json').urls;
 const utils = require('../utils');
 
-describe("Browser", function() {
+describe("Select", function() {
     this.timeout(5000);
     let browser;