+++ tests/browser_base/query_test.js
@@ -1,0 +1,62 @@
+"use strict";
+
+const Wendigo = require('../../lib/wendigo');
+const assert = require('assert');
+const configUrls = require('../config.json').urls;
+const utils = require('../utils');
+
+describe("Query", function() {
+    this.timeout(5000);
+
+    let browser;
+    before(async () => {
+        browser = await Wendigo.createBrowser();
+    });
+
+    beforeEach(async () => {
+        await browser.open(configUrls.index);
+    });
+
+    after(async () => {
+        await browser.close();
+    });
+
+    it("Query", async () => {
+        const element = await browser.query("h1");
+        assert(element);
+    });
+
+    it("Query Multiple Elements", async () => {
+        const element = await browser.query("p");
+        assert(element);
+    });
+
+    it("Query Not Element Found", async () => {
+        const element = await browser.query("div.not-exists");
+        assert.strictEqual(element, null);
+    });
+
+    it("Query Node", async () => {
+        const element = await browser.query("h1");
+        const element2 = await browser.query(element);
+        assert(element2);
+    });
+
+    it("Query Sub Element", async () => {
+        const container = await browser.query(".container");
+        const element = await browser.query(container, "p");
+        assert(element);
+    });
+
+    it("Query Sub Element Not Found", async () => {
+        const container = await browser.query(".container");
+        const element = await browser.query(container, "b");
+        assert.strictEqual(element, null);
+    });
+
+    it("Query Sub Element Not Valid Parent", async () => {
+        utils.assertThrowsAsync(async () => {
+            await browser.query(".container", "b");
+        }, "Error: Invalid parent element for query");
+    });
+});