+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
   * Added browser.innerHtml and assertions
+  * Query and QueryAll support optionally 2 parameters for subqueries
   * Assert text and not text now support arrays of expected strings
   * Minor tests cleanup